AIAgent/requirements.txt
fastapi
uvicorn
pandas
numpy
matplotlib
sentence-transformers
requests
streamlit
python-dotenv
prophet
openai
aiogram
lightgbm
python-multipart
plotly
statsmodels
smolagents[litellm]
accelerate
redis
datasets
uuid
scikit-learn
catboost
xgboost
scipy
kagglehub